# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for fetching product item",
        "title": "Product API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Alex Duzi",
            "email": "duzihd@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/products": {
            "get": {
                "description": "Get a list of all products with thumbnails",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "List all products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ProductListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "description": "Get product details by product ID including all images",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Get a product by ID",
                "parameters": [
                    {
                        "type": "string",
                        "example": "MLB001",
                        "description": "Product ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.ProductResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/errors.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns the health status of the API",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check endpoint",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.HealthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "errors.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "PRODUCT_NOT_FOUND"
                },
                "error": {
                    "type": "string",
                    "example": "product not found"
                },
                "message": {
                    "type": "string",
                    "example": "The requested product does not exist"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "handler.HealthResponse": {
            "type": "object",
            "properties": {
                "service": {
                    "type": "string",
                    "example": "product-api"
                },
                "status": {
                    "type": "string",
                    "example": "healthy"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "handler.ProductListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.ProductDTO"
                    }
                }
            }
        },
        "handler.ProductResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/usecase.ProductDTO"
                }
            }
        },
        "usecase.ProductDTO": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "Electronics \u003e Smartphones"
                },
                "condition": {
                    "type": "string",
                    "example": "new"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                },
                "currency": {
                    "type": "string",
                    "example": "USD"
                },
                "description": {
                    "type": "string",
                    "example": "Latest Apple flagship smartphone with A17 Pro chip"
                },
                "id": {
                    "type": "string",
                    "example": "MLB001"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/usecase.ProductImageDTO"
                    }
                },
                "price": {
                    "type": "number",
                    "example": 1299.99
                },
                "seller_id": {
                    "type": "string",
                    "example": "SELLER001"
                },
                "seller_name": {
                    "type": "string",
                    "example": "TechWorld Store"
                },
                "stock": {
                    "type": "integer",
                    "example": 45
                },
                "thumbnail": {
                    "type": "string",
                    "example": "https://images.unsplash.com/photo-1696446702230-a8ff49103cd1?w=800"
                },
                "title": {
                    "type": "string",
                    "example": "iPhone 15 Pro Max 256GB - Titanium Blue"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-01T00:00:00Z"
                }
            }
        },
        "usecase.ProductImageDTO": {
            "type": "object",
            "properties": {
                "display_order": {
                    "type": "integer",
                    "example": 0
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "image_url": {
                    "type": "string",
                    "example": "https://images.unsplash.com/photo-1696446702230-a8ff49103cd1?w=800"
                },
                "product_id": {
                    "type": "string",
                    "example": "MLB001"
                }
            }
        }
    }
}